Moved renamePlaylist to Homework21
moved rename playlist to homework
refactored Profile test (got rid of Thread)
