+++ src/test/java/HomeTest.java
@@ -11,7 +11,6 @@ import java.util.List;
 
 public class HomeTest extends BaseTest{
 
-    String newPlaylistName = "Sample Edited Playlist";
 
     @Test
     public void HoverOverPlayButtonAndPlaySong(){
@@ -38,40 +37,10 @@ public class HomeTest extends BaseTest{
         //Assert returns String 1.contains(String 2)
     }
 
-    @Test
-    public void renamePlaylist() {
-        String updatedPlaylistMessage = "Updated playlist \"Sample Edited Playlist.\"";
 
-        provideEmail("sviatlana.rysiavets@testpro.io");
-        providePassword("nTtAZKUq");
-        clickSubmit();
-        //double click on playlist name
-        doubleClickPlaylist();
-        //enter new Name
-        enterNewName();
-        Assert.assertEquals(getRenamePlaylistSuccessMsg(), updatedPlaylistMessage);
-
-    }
 
     //Helper Methods
-    public String getRenamePlaylistSuccessMsg(){
-        WebElement notification = wait.until(ExpectedConditions.visibilityOfElementLocated(By.cssSelector("div.success.show")));
-        return notification.getText();
-    }
 
-    public void enterNewName() {
-        WebElement playlistInputField = wait.until(ExpectedConditions.visibilityOfElementLocated(By.cssSelector("[name='name']")));
-        //clear() does not work, element has an attribute of 'required'
-        //workaround is Ctrl A (to select all) then backspace to clear then replace with new playlist name
-        playlistInputField.sendKeys(Keys.chord(Keys.COMMAND, "A", Keys.BACK_SPACE));
-
-        playlistInputField.sendKeys(newPlaylistName);
-        playlistInputField.sendKeys(Keys.ENTER);
-    }
-    public void doubleClickPlaylist(){
-        WebElement playlist = wait.until(ExpectedConditions.visibilityOfElementLocated(By.cssSelector(".playlist:nth-child(3)")));
-        actions.doubleClick(playlist).perform();
-    }
 
     public String getPlaylistDetails(){
         return driver.findElement(By.cssSelector("span.meta.text-secondary span.meta")).getText();
