+++ src/test/resources/TestNG.xml
@@ -13,6 +13,7 @@
             <class name="Homework18"/>
             <class name="CreatePlaylist"/>
             <class name="Homework17"/>
+            <class name="Homework21"/>
             <class name="Homework20aka19"/>
             <class name="AllSongsTest"/>
             <class name="HomeTest"/>
